% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{change_content_types}
\alias{change_content_types}
\title{Change content types}
\usage{
change_content_types(metadata, board)
}
\arguments{
\item{metadata}{Metadata of element}

\item{board}{`dspins_board_s3` board}
}
\description{
Change content types
}
