% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/links.R
\name{create_ds_links}
\alias{create_ds_links}
\title{Create ds links}
\usage{
create_ds_links(slug, folder, bucket, formats, element_type)
}
\arguments{
\item{slug}{Slug of element}

\item{folder}{Folder of user}

\item{bucket}{AWS S3 bucket}

\item{formats}{File formats}

\item{element_type}{Element type ('dsviz', 'fringe', or 'drop')}
}
\value{
List containing a list of links to files in ds profile and a list of links to share element
}
\description{
Create list of links to files in ds profile and to share element.
}
