% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boards.R
\name{ds_board_s3}
\alias{ds_board_s3}
\title{Use a folder within an AWS S3 bucket as a board.}
\usage{
ds_board_s3(
  user_name = NULL,
  org_name = NULL,
  bucket_id = NULL,
  versioned = FALSE,
  access_key = NULL,
  secret_access_key = NULL,
  session_token = NULL,
  credential_expiration = NULL,
  profile = NULL,
  region = NULL,
  endpoint = NULL,
  cache = NULL
)
}
\arguments{
\item{user_name}{DS user name}

\item{org_name}{DS org name}

\item{bucket_id}{DS bucket id. Defaults to `bucket_id = "user"`}
}
\description{
For authentication, an `.env` file with the necessary `AWS_ACCESS_KEY_ID`
and `AWS_SECRET_ACCESS_KEY` env vars needs to be located in the root directory.
}
\details{
If both `user_name` and `org_name` are specified, `org_name` is used.
}
\examples{
\dontrun{
board <- ds_board_s3(user_name = "test", bucket_id = "user")
}
}
