% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dspin-exist-delete-meta.R
\name{dspin_delete}
\alias{dspin_delete}
\title{Delete DS pins}
\usage{
dspin_delete(board, names, ...)
}
\arguments{
\item{board}{`dspins_board_s3` board}

\item{...}{}

\item{name}{Name of element to be deleted}
}
\description{
Delete a DS pin from `dspins_board_s3`.
}
\examples{
\dontrun{
board <- ds_board_s3(user_name = "test", bucket_id = "user")

board \%>\% dspin_delete("mtcars-dataset")
}
}
