% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dspin-download.R
\name{dspin_download}
\alias{dspin_download}
\title{Download DS pins}
\usage{
dspin_download(board, name, hash = NULL, ...)
}
\arguments{
\item{board}{`dspins_board_s3` board}

\item{name}{Name of element to be read}

\item{...}{}
}
\value{
File paths for downloaded files
}
\description{
Download a DS pin to locale cache.
}
\examples{
\dontrun{
board <- ds_board_s3(user_name = "test", bucket_id = "user")

board \%>\% dspin_download("slug-of-drop-pin")
}
}
