% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dspin-exist-delete-meta.R
\name{dspin_meta}
\alias{dspin_meta}
\title{DS pin meta}
\usage{
dspin_meta(board, name, version = NULL, ...)
}
\arguments{
\item{board}{`dspins_board_s3` board}

\item{name}{Name of element}

\item{...}{}
}
\value{
List of metadata
}
\description{
Get meta data of DS pin in board of type `dspins_board_s3`.
}
\examples{
\dontrun{
board <- ds_board_s3(user_name = "test", bucket_id = "user")

board \%>\% dspin_meta("mtcars-dataset")
}
}
