% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dspin-read-write.R
\name{dspin_read}
\alias{dspin_read}
\title{Read DS pins}
\usage{
dspin_read(board, name, hash = NULL, ...)
}
\arguments{
\item{board}{`dspins_board_s3` board}

\item{name}{Name of element to be read}

\item{...}{}
}
\value{
DS element
}
\description{
Read a `fringe` or `dsviz` element to a board of type `dspins_board_s3`.
}
\details{
Raises error for element of type `drop`.
}
\examples{
\dontrun{
board <- ds_board_s3(user_name = "test", bucket_id = "user")

board \%>\% dspin_read("mtcars-dataset")
}
}
