% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save-fringe.R
\name{dspin_save}
\alias{dspin_save}
\title{Save DS elements}
\usage{
dspin_save(element, slug, board, path, ...)
}
\arguments{
\item{element}{Element to be saved (`fringe`, `dsviz`, or `drop`)}

\item{slug}{Slug of element to be saved}

\item{board}{`dspins_board_s3` board}

\item{path}{Path to save elements to}

\item{...}{}
}
\value{
Metadata for element
}
\description{
Preprocesses metadata for elements of type `fringe`, `dsviz`, or `drop`
and saves them to given path.
}
