% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dspin-read-write.R
\name{dspin_write}
\alias{dspin_write}
\title{Write DS pins}
\usage{
dspin_write(board, element, ...)
}
\arguments{
\item{board}{`dspins_board_s3` board}

\item{element}{Element to be saved (`fringe`, `dsviz`, or `drop`)}

\item{...}{}
}
\description{
Pin a `fringe`, `dsviz`, or `drop` element to a board of type `dspins_board_s3`.
}
\examples{
\dontrun{
board <- ds_board_s3(user_name = "test", bucket_id = "user")

fringe_mtcars <- homodatum::fringe(mtcars, name = "Mtcars dataset")
board \%>\% dspin_write(fringe_mtcars)
}
}
