% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dspin-exist-delete-meta.R
\name{pin_delete}
\alias{pin_delete}
\title{Delete DS pins}
\usage{
pin_delete(board, name, ...)
}
\arguments{
\item{board}{`dspins_board_s3` board}

\item{name}{Name of element to be deleted}

\item{...}{}
}
\description{
Delete a DS pin from `dspins_board_s3`.
}
\examples{
\dontrun{
board <- ds_board_s3(user_name = "test", bucket_id = "user")

board \%>\% pin_deleted("mtcars-dataset")
}
}
