% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dspin-exist-delete-meta.R
\name{pin_exists.dspins_board_s3}
\alias{pin_exists.dspins_board_s3}
\title{DS pin exists}
\usage{
\method{pin_exists}{dspins_board_s3}(board, name, ...)
}
\arguments{
\item{board}{`dspins_board_s3` board}

\item{name}{Name of element to be checked}

\item{...}{}
}
\value{
Boolean; `TRUE` if DS pin exists
}
\description{
Check for existence of DS pin in board of type `dspins_board_s3`.
}
\examples{
\dontrun{
board <- ds_board_s3(user_name = "test", bucket_id = "user")

board \%>\% pin_exists("mtcars-dataset")
}
}
