% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dspin-store-fetch.R
\name{pin_fetch.dspins_board_s3}
\alias{pin_fetch.dspins_board_s3}
\title{Fetch a DS pin}
\usage{
\method{pin_fetch}{dspins_board_s3}(board, slug, ...)
}
\arguments{
\item{board}{`dspins_board_s3` board}

\item{slug}{Slug of element to be downloaded}
}
\value{
Metadata of element
}
\description{
Used by `pin_read()` to download pins on `dspins_board_s3` to local cache.
}
