% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dspin-list.R
\name{pin_list}
\alias{pin_list}
\title{DS pin list}
\usage{
pin_list(board, name, ...)
}
\arguments{
\item{board}{`dspins_board_s3` board}

\item{...}{}

\item{extended}{Boolean to get simple or extended list, defaults to `FALSE`}
}
\value{
If `extended = FALSE`, returns character vector of names of DS pins in board;
if `extended = TRUE`, returns `data.frame` showing contents of the board's `data.txt`.
}
\description{
Get list of all pins saved to a board of type `dspins_board_s3`.
}
\examples{
\dontrun{
board <- ds_board_s3(user_name = "test", bucket_id = "user")

board \%>\% pin_list()
board \%>\% pin_list(extended = TRUE)
}
}
