% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dspin-store-fetch.R
\name{pin_store}
\alias{pin_store}
\title{Store a DS pin}
\usage{
pin_store(board, element, ...)
}
\arguments{
\item{board}{`dspins_board_s3` board}

\item{slug}{Slug of element to be downloaded}

\item{path}{Path to which DS pins have been saved}

\item{metadata}{Metadata of DS pin}
}
\value{
Board
}
\description{
Used by `pin_write()` to upload local DS file(s) to `dspins_board_s3`,
not to be called directly.
}
