% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dspin-urls.R
\name{pin_urls}
\alias{pin_urls}
\title{Write DS pins and return URLs}
\usage{
pin_urls(board, element = NULL, ...)
}
\arguments{
\item{board}{`dspins_board_s3` board}

\item{element}{Element to be saved (`fringe`, `dsviz`, or `drop`)}

\item{...}{}
}
\value{
List of links to pin in DS profile
}
\description{
Pin a `fringe`, `dsviz`, or `drop` element to a board of type `dspins_board_s3`
and return links to pin on DS profile.
}
\examples{
\dontrun{
board <- ds_board_s3(user_name = "test", bucket_id = "user")

fringe_mtcars <- homodatum::fringe(mtcars, name = "Mtcars dataset")
dsurls <- board \%>\% pin_urls(fringe_mtcars)
}
}
