% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-datatxt.R
\name{update_datatxt}
\alias{update_datatxt}
\title{Update datatxt}
\usage{
update_datatxt(metadata, board)
}
\arguments{
\item{metadata}{List of metadata of pin being saved.}

\item{board}{Pin board of type `dspins_board_s3`.}
}
\description{
Downloads the `data.txt` file for a given board and updates metadata.
}
\details{
If the an entry already exists for the `slug` passed through `metadata$slug`,
the existing entry is overwritten; otherwise it is added.
}
