% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dspins-package.R
\docType{package}
\name{dspins-package}
\alias{dspins}
\alias{dspins-package}
\title{dspins: DS pins}
\description{
Save DS files.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/datasketch/dspins}
  \item Report bugs at \url{https://github.com/datasketch/dspins/issues}
}

}
\author{
\strong{Maintainer}: Juan Pablo Marin Diaz \email{jpmarindiaz@gmail.com}

}
\keyword{internal}
